refactor: creation of the searchBar.js file and remotion of the search bar logic from the query.js file
